+++ src/components/StudyMain/StudyComponents.tsx
@@ -8,8 +8,8 @@ import { Member } from './Member';
 import useSWR from 'swr';
 import { useRouter } from 'next/router';
 import useUser from 'src/libs/client/useUser';
-import { StudyCalendarWrap } from '../Calendar/StudyPage/StudyCalendarWrap';
 import { IStudyResponse } from 'src/types/Study';
+import { StudyCalendarWrap } from '../Calendar/StudyPage/StudyCalendarWrap';
 
 export const StudyComponents = () => {
   const router = useRouter();
@@ -33,13 +33,13 @@ export const StudyComponents = () => {
     <>
       {myStudy ? (
         <Container>
-          <Record />
+          {/* <Record />
           <TodoList studyId={data?.study?.id} />
           <Notice studyId={data?.study?.id} />
           <Member
             memberslimit={data?.study?.membersLimit}
             memberlist={data?.study?.joinMember}
-          />
+          /> */}
           <StudyCalendarWrap />
         </Container>
       ) : (
